Update README.md
This application is a university specific application that allows the user to stay connected with nightlife and friends from their mobile device. The application will turn your location services on, and detect which bars you enter on campus. Once a user enters the vicinity of a bar, a drink special/food special menu will pop up.